% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/generate.R
\name{generateDrugCohort}
\alias{generateDrugCohort}
\title{Generate a Drug Cohort}
\usage{
generateDrugCohort(
  connectionDetails,
  cdmDatabaseSchema,
  resultsDatabaseSchema,
  drugConceptIds = c()
)
}
\arguments{
\item{connectionDetails}{An R object of type\cr\code{connectionDetails} created using the function
\code{createConnectionDetails} in the \code{DatabaseConnector} package.}

\item{cdmDatabaseSchema}{The fully qualified name of the CDM schema}

\item{resultsDatabaseSchema}{The fully qualified name of the Results schema to store the cohort}

\item{drugConceptIds}{A list of drug concepts}
}
\description{
Generate a Drug Cohort
}
