% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/generate.R
\name{getCharlsonForDrugCohort}
\alias{getCharlsonForDrugCohort}
\title{Get Charlson Index for a Drug cohort}
\usage{
getCharlsonForDrugCohort(
  connectionDetails,
  cdmDatabaseSchema,
  cohortName = "drugCohort",
  pathToSqlFile = "",
  drugConceptIds = c(),
  sqlOnly = FALSE
)
}
\arguments{
\item{connectionDetails}{An R object of type\cr\code{connectionDetails} created using the function
\code{createConnectionDetails} in the \code{DatabaseConnector} package.}

\item{cdmDatabaseSchema}{The fully qualified name of the CDM schema}

\item{cohortName}{A readable name of the drug cohort, used for writing out the SQL in SqlOnly mode}

\item{pathToSqlFile}{The path to use to write out the SQL file, when using SqlOnly mode}

\item{drugConceptIds}{A list of drug concepts}

\item{sqlOnly}{Should we just export a SQL file, or actually run the query?}
}
\description{
Get Charlson Index for a Drug cohort
}
